Moved user_scenarios.py from scripts directory to sovrin_node.utils package.
