+++ scripts/test_users_write_and_read_own_keys.py
@@ -20,8 +20,8 @@ from concurrent.futures import ProcessPoolExecutor
 
 from stp_core.common.log import getlogger
 
-from scripts.user_scenarios import generateNymsData, NymsCreationScenario, \
-    KeyRotationAndReadScenario
+from sovrin_node.utils.user_scenarios import generateNymsData, \
+    NymsCreationScenario, KeyRotationAndReadScenario
 
 STEWARD1_SEED = b"000000000000000000000000Steward1"
 
@@ -61,7 +61,7 @@ def main(args):
 
     with ProcessPoolExecutor(numOfUsers) as executor:
         usersIdsAndVerkeys = [(user.identifier, user.verkey)
-                               for user in users]
+                              for user in users]
 
         nymsCreationScenarioFuture = \
             executor.submit(NymsCreationScenario.runInstance,
